Revert "Refactor/pick a faster json lib (#1152)"
This reverts commit c3e5747cb4bcf633a35d6591a1b0503b64a89d1e.
