+++ examples/ExecutionPools/Discover/test_plan.py
@@ -45,7 +45,6 @@ def main(plan):
 
 if __name__ == "__main__":
     res = main()
-    if res.report.entries:
-        assert len(res.report.entries) == 5
+    assert len(res.report.entries) == 5
     print("Exiting code: {}".format(res.exit_code))
     sys.exit(res.exit_code)
