+++ testplan/web_ui/testing/src/AssertionPane/AssertionTypes/basicAssertionUtils.js
@@ -35,18 +35,27 @@ function prepareLogContent(assertion, defaultContent) {
   let decodedMsg = null;
 
   if (assertion.message !== undefined) {
-    decodedMsg = (
-      <Linkify
-        options={{
-          target: "_blank",
-          validate: {
-            url: (value) => /^https?:\/\//.test(value),
-          },
-        }}
-      >
-        {assertion.message}
-      </Linkify>
-    );
+    let bytearray;
+    if (
+      typeof assertion.message === "object" &&
+      typeof (bytearray = assertion.message["_BYTES_KEY"]) !== "undefined" &&
+      Array.isArray(bytearray)
+    ) {
+      decodedMsg = bytearray.length ? String.fromCodePoint(...bytearray) : "";
+    } else {
+      decodedMsg = (
+        <Linkify
+          options={{
+            target: "_blank",
+            validate: {
+              url: (value) => /^https?:\/\//.test(value),
+            },
+          }}
+        >
+          {assertion.message}
+        </Linkify>
+      );
+    }
   }
 
   const preContent = <pre>{decodedMsg}</pre>;
@@ -237,18 +246,27 @@ function prepareFailContent(assertion, defaultContent) {
   let decodedMsg = null;
 
   if (assertion.message !== undefined) {
-    decodedMsg = (
-      <Linkify
-        options={{
-          target: "_blank",
-          validate: {
-            url: (value) => /^https?:\/\//.test(value),
-          },
-        }}
-      >
-        {assertion.message}
-      </Linkify>
-    );
+    let bytearray;
+    if (
+      typeof assertion.message === "object" &&
+      typeof (bytearray = assertion.message["_BYTES_KEY"]) !== "undefined" &&
+      Array.isArray(bytearray)
+    ) {
+      decodedMsg = bytearray.length ? String.fromCodePoint(...bytearray) : "";
+    } else {
+      decodedMsg = (
+        <Linkify
+          options={{
+            target: "_blank",
+            validate: {
+              url: (value) => /^https?:\/\//.test(value),
+            },
+          }}
+        >
+          {assertion.message}
+        </Linkify>
+      );
+    }
   }
 
   const preContent = <pre>{decodedMsg}</pre>;
